consolidate Installation back into single index.ts file
